fix(ci): green the pre-existing red CI from 9f130cf0 (fmt, clippy -D, stale API snapshots)
CI has been red since the zq-seed commit (9f130cf0) on three jobs, none
related to the lossless work:

- Format: zq_seed.rs + zq_target.rs were pushed un-fmt'd (already cured
  by the fmt ride-along in 6226444c; cargo fmt --check now clean).
- Clippy -D warnings: the current stable clippy promotes
  chunks_exact-with-constant-size to a lint; fixed all sites via
  as_chunks across lib/tests/benches in both CI feature configurations
  (default and parallel,moxcms,ultrahdr,zencodec,__test-utils — both now
  0 errors). Also: removed genuinely-unused imports in
  target_quality_real.rs; arch-gated x86-test-only helpers
  (decode_path_dispatch_parity.rs, ycbcr.rs turbo_ref_rgb,
  simd_types.rs quantize_test_data) that are dead on aarch64; zenyuv:
  gated pack_i16_pair + rgb_to_yuv420_scalar_tail to the x86_64/wasm32
  kernels that call them and dropped a redundant i32 cast in the NEON
  kernel (aarch64-only file, no behavior change).
- Public API surface: regenerated docs/public-api snapshots with the
  pinned nightly-2026-07-13 via just api-doc. The surface growth is the
  zq-seed commit's target_quality (17 items) + zq_seed (2 items)
  modules; the lossless fixes added zero public items.

Verified locally: both CI clippy invocations 0 errors; cargo fmt --check
clean; apidoc public_api_surface_docs_are_current passes; lib suite
1093 tests + lossless matrix + dispatch parity all green.

+++ zenyuv/src/neon_encode.rs
@@ -126,8 +126,8 @@ fn compute_channel_neon(
     );
 
     // Shift right by PREC and narrow to i16.
-    let shifted_lo = vshrn_n_s32::<{ crate::types::PREC as i32 }>(sum_lo);
-    let shifted_hi = vshrn_n_s32::<{ crate::types::PREC as i32 }>(sum_hi);
+    let shifted_lo = vshrn_n_s32::<{ crate::types::PREC }>(sum_lo);
+    let shifted_hi = vshrn_n_s32::<{ crate::types::PREC }>(sum_hi);
 
     vcombine_s16(shifted_lo, shifted_hi)
 }
